pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grpo-composer"
version = "0.1.0"
description = "A Unified, Component-Driven Library for Critic-Free Reinforcement Learning in LLMs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "GRPO-Composer Team"}
]
keywords = ["reinforcement-learning", "llm", "grpo", "rlhf", "transformers"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "datasets>=2.14.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
verl = [
    "verl==0.6.1",
    "ray[default]>=2.40.0",
    "vllm==0.8.5",
    "transformers>=4.51.0,<5.0.0",
    "tensordict>=0.8.0,<=0.10.0,!=0.9.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/grpo-composer/grpo-composer"
Documentation = "https://grpo-composer.readthedocs.io"

[tool.setuptools.packages.find]
where = ["."]
include = ["grpo_composer*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
